Minor Gradle Property Tweaks. (#2367)
Expose the workers.max property in gradle.properties. This allows a dev to set the number of processors Gradle uses per their needs. See this.

Also take the chance to turn on vfs.watch. This allows Gradle to watch the virtual file system it creates in memory. In theory, this makes it slightly faster for incremental builds. See this.