Organise and label DefaultJobPersistence tests. (#2388)
I took this opportunity to reorganize the tests with JUnit 5's nested class. I also added more human friendly descriptions to each test method via the DisplayName annotation.

In general, I've nested tests for the same method under the same class (usually named after the method being tested). The classes are alphabetically sorted for easy reading. I've left singular tests at the top of the file (they are easy to read as is). This should make it easier for devs to jump to the methods they are interested in, read the related test cases to understand behavior, and write new tests.

I also added test cases to verify the list methods are returning Jobs in descending order and an additional enqueue test to clarify enqueue job behaviour.