Implement majority of more efficient Redshift Copy Strategy. (#2547)
Instead of inserts, we write the data to S3 and issue a COPY command copying from S3 into redshift.

Use a single file as its sufficiently performant and we do not want to introduce file destination related operations yet.
Use an open source library for uploads as AWS does not natively support streaming loads.

My intention with this PR is to first implement the meat of the write-and-copy strategy. This is mainly centered around the RedshiftCopier class. I plan hook up the RedshiftCopier to the actual Destination class, and implement all the plumbing, in a follow up PR.

Co-authored-by: Davin Chia <davinchia@Davins-MacBook-Pro.local>