:tada: Introduce Redshift Destination COPY Strategy (#2606)
Plumb everything together to allow a user to, via configuration, pick either the less efficient and easier to set up INSERT strategy, or the more efficient and slightly more set up COPY strategy.

Effectiveness of COPY is obvious with larger and larger number of rows for a rough 3 - 4 x improvement. Logs show we are now effectively read constrained.

Modify the JSON spec to take additional S3 configuration. The RedshiftDestination is split into the InsertDestination and the CopyDestination. The RedshiftDestination picks the CopyDestination if this S3 configuration exists, falling back to the InsertDestination otherwise.

The original RedshiftDestination is now the RedshiftInsertDestination class - all that previous functionality is identical.