AcceptTracked should accept an AirbyteRecordMessage. (#2662)
The acceptTracked method should accept an AirbyteRecordMessage instead of a generic AirbyteMessage. This allows us to centralise checking for a record and makes the interface easier to understand.

We can also consolidate checking if a received message has a corresponding stream. However that's more involved and can be revisited at a later date.