Update stripe data types to number instead of int
For some reason a lot of integer fields are being output as `<value>.0` which breaks normalization. This is also happening in other sources, Shopify and Recurly. We need to root cause this issue or roll out a wider fix, perhaps using the CDK. For now, we're just trying to unblock the user.