Clean up Cancel Exception Handling. (#3891)
As part of working on the Kube cancelling, I realised our exception handling for the cancel operation is noisy and confusing. It was quite scary to me as I tried to debug why Kube was not happening. Decided to clean this up as is before nailing down Kube cancel behaviour.

This PR tries to clean this up. Unnecessary exceptions have been removed. Exceptions that happen on cancel have been converted into log lines that state this exception is typically thrown on cancel.

In particular,
- Convert IOException in LineGobbler to log lines since this a race occurs between closing the pool and the process. An exception must be thrown here when the gobbler tries to read the stream after the underlying process has exited.
- Stop catching and throwing exceptions in the CancellationHandler. This exception is also thrown as this is how Temporal handles cancellation. Convert this to a log line.

See the PR for more detailed notes.