Better logging for Job Submission errors. (#4452)
Log the job id, job status and connection id if there is any job submission error to allow devs to easily investigate what is going on.

Without this it's impossible to remove the offending job. This is bad since the job submitter is currently set up to pull the oldest job, which means it will spin on the offending job, stopping all job submission.