Fix Kube S3 Logging Bug. (#4914)
See bluedenim/log4j-s3-search#101 for a summary.

Because this is upstream, I built the jar locally and uploaded it to a temp gcs bucket serving as a maven repository.

I've added this location and the new version in the build.gradle.

I want to merge this in now since there are several users waiting on this. Created #4915 to follow up with the OSS PR and clean up work.