Use Cloudrepo.io to host temporary built jars. (#4949)
Continuation of my saga to publish a public jar while we wait for the upstream fix to be merged in.

Turns out using the GCS bucket as a public maven repository still requires GCS login. This is not feasible since it requires all our community contributors to have and sign into gcloud. This also required injecting this everywhere in the build and was extremely ugly.

We could publish to the central maven repository, but it takes 2 days for us to get an OSS account. We'd also have to dedup on artifact id.
I explored GCP's Artifact Registry, but I didn't manage to set it up after an hour. It is also only possible to upload via Maven/Gradle and there were a lot of steps to jump through to authenticate.
I went with CloudRepo as they were the best in term of set up, UX and performance. I was able to set something up in 10 mins. I was able to upload my Jars manually via the UI. CoudRepo also supports public read-only repository, which is exactly what we need (publish is still gated on authentication). Ran into a slight hiccup and support got back to me in 30 mins. Frankly think we should consider using them for our private cloud. They are also fairly cheap. Anyone who wants to upload jars will have to use credentials that I've added to shared-engineering folder. Just navigate to the CloudRepo url.