Bug Fix: Handle missing data types in migrations. (#5089)
Previously, when Airbyte exported configurations, even if there were not configurations of a type (e.g. operations) it would create a placeholder file for it in the export archive. The implementation has changed such that we cannot make that guarantee anymore. Specifically the postgres database only exports data that it actually has.

The migration assumed that every archive it processed would contain data from all types. This commit makes the change to make sure that the types represented in the archive just need to be a subset of the types declared in the schema.

In the future we can decide if we have an opinion on whether we should always export placeholder files. Since we have versions of Airbyte that don't do it though, we need to make this change in the migration code.