Move podSweeper to recreate strategy
Due to the volume mounts, it needs to be recreated so volumes can be moved to a new node. Potentially, we could add an affinity so that the volume is already attached to the node for no downtime. Another alternative would be to use object storage or something for this.