Set up initial metrics reporter. (#10408)
Part 1 of https://docs.google.com/document/d/11pEUsHyKUhh4CtV3aReau3SUG-ncEvy6ROJRVln6YB4/edit?usp=sharing.

This is the initial set up of the reporter. Will add actual metrics in the follow up PR.

Since the reporter is primarily for Cloud use and limited to DD to begin with, we are keeping it outside the regular Airbyte docker and Kube deploys for now.

Add ReadMes to better define various modules.