Turn on zombie build instance termination. (#11098)
Script to remove zombie build instances to prevent resource wastage. Consider a build instance that has been running for more than an hour a zombie instance as the average runs is ~20 mins and the most I've seen this go is 35 mins.

Required since:
- We cannot guarantee our actions correctly terminate instances.
- Failsafe against a dev accidentally writing a workflow that doesn't terminate instances.

First bit to identify such instances was merged in a day ago for testing. This is a follow up now that I've seen a day worth of runs and confirms the query works.