Implement health check for server. (#11755)
As part of digging into airbytehq/oncall#192. I realised the OSS Server's health checks aren't actually implemented. This was removed in ##9288. However we do require this to be implemented since we use it in Airbyte Cloud to inform whether the server is connected to the database or not and thus have properly start up.

Since this check can potentially be called multiple times a second, it is important for the check to be fast. Because of this, we check the server is working by attempting to list the id column of the workspace table's first row.