Restore jsonPath and fix it (#12325)
This restore the Json traversal library. A bug was introduce in the Json path library, the PR fix it.

In a json schema we can define an enum without specifying a "type" attribute. It wasn't handle in the previous implemantation. We now return a right type in the getType method and process it the same way than the an integer/boolean/string type.