🎉  Inject Job Type into processes names. (#12561)
We currently use sync for all process names (e.g. container names on Docker and pod names on Kube). This is confusing as though syncs are the majority of operations, not all operations are syncs.

This PR makes job type part of the process name so it's clearer to airbyte operators what processes correspond to what jobs. Intention is to help with oncall.

After this change, process names will look like (truncated to fit the underlying system):

- check = <connector-image>-check-<job-info>-<random-suffix>
- spec = <connector-image>-spec-<job-info>-<random-suffix>
- discover = <connector-image>-discover-<job-info>-<random-suffix>
- sync = <connector-image>-read-<job-info>-<random-suffix> for source connectors, <connector-image>-write-<job-info>-<random-suffix> for destination connectors, and <connector-image>-normalize-<job-info>-<random-suffix> for normalization processes.
- custom operations = <connector-image>-custom-<job-info>-<random-suffix> (not enabled on Cloud).