Support the ssh tunnel port option in workers
In Airbyte Cloud, the SSH tunnel port option is not used when using a
port different than 22.

This adds support for the tunnel_port option to the sshtunneling.sh script.

This patch is untested, because I don't have time to setup a local test
environment (this is the reason to use cloud in the first place).

A more detailed bug report can be found at #13758.