:bug: fix connect err when setting off pg bitnami chart
Current status of chart would only set POSTGRES_SEEDS to the default bitnami pg deployed by default. However for prod it was not possible to use an external host database. Now chart takes the database host set in global, if none is set it defaults to default bitnami pg host deployed with the chart.