🪟 🔧 Add PostHog A/B Testing tool
Posthog is an analytics tool that will allow us to run A/B tests in
product in a simple manner.

The advantage vs LaunchDarkly is that they provide features for
experimenting, setting goals, create funnels, etc. Thanks to their UI, we can define experiments and
monitor which variant is the winner. Posthog will do all the maths for
us.
More info: https://posthog.com/product/experimentation-suite

This addition needs to be approved by @timroes after a security audit
has been performed.

This PR sets all the code needed in case the audit is succesful