More Readable Gradle test config. (#18890)
The integration testing logs are extremely noisy today. The logs contain the entire test STDOUT and are a pain to read.

This PR contains a better test setup config to reduce logging output.

The main change is to comment out ShowStandardStreams in the integration tests plugin.
Then we add the started event so Gradle shows us when a test is started. Some tests take a longer time. It is useful to know they have begun so it's clear if the job is stuck or not.
Standardise the above setting across all of our custom test tasks.