Job Persistence Refactor in preparation for progress bar. (#18891)
Some refactoring in preparation for the progress bar persistence changes.

The main change here was to simplify some of the JobPersistence methods by moving the logic to calculate attemptId into the JobPersistence implementation. This logic currently sits outside the class and is duplicated in multiple places. We could expose a helper method to calculate this logic, however that felt unnecessary at this point.

The alternative is further duplicating this logic as the progress bar logic is implemented, so I want to get that out of the way.

The other reason it's cleaner to use jobId and attemptNumber is these concepts/terms are more familiar throughout the rest of the codebase and it feels more intuitive to continue speaking this language (in my opinion).

Some random bits I wanted to clean up on the way as well. I will leave comments in the files as appropriate.