airbyte-to-flow: Plumb `--log.level` flag to `LOG_LEVEL` env var
This is, I believe, the last missing piece before we're able to
modify the log level in a shard spec (`flowctl-admin shards edit`)
and have that take effect in the connector itself.

However, while the connectors we've written all support the flag
`--log.level` we probably don't want to assume that any arbitary
connector supports it. So instead of providing the underlying
connector entrypoint with a `--log.level` flag we instead set
the `LOG_LEVEL` environment variable on the subprocess execution.

Since our boilerplate package `flow/protocols/airbyte` already
supports environment variables as well as flags this should work
as an opt-in mechanism without breaking anything that doesn't
currently support it.