🪟 🐛 Fix custom connection creation endpoint (#19702)
Closes #19637

PR that introduced this bug was a precision revert of another issue: #19627

This PR removes the createSourceDefinition() and createDestinationDefinition() methods in favor of the new createCustomSourceDefinition() and createCustomDestinationDefinition() methods. createSourceDefinition() and createDestinationDefinition() are to be deprecated.

This PR does not address updating custom connectors. There is a backend issue for fixing this behavior here: #19669