Stop `cursor` pagination when stream state value reached
Currently the cursor pagination specially for `TicketAudits` keeps paginating
until the stream finishes every time without taking into consideration neither the
`start_time` set in the configuration nor the `stream_state` value. It is expected
to stop once it reaches the `start_time` or the previous value of the stream state.