Add orchestrator label. (#20904)
Add the orchestrator label to orchestrators so we can better differentiate orchestrator pods.

This is useful since orchestrator pods are the only pods in the job namespace with a need to talk to the main Airbyte application pods. These labels allow us to apply more granular network filtering.

Also took the chance to do some clean up of labels.