Microsoft has migrated from ADAL4j to MSAL4J
Attempting to use the source-mssql connector with ActiveDirectoryPassword (or other ActiveDirectory parameters in the connection string) cause an error "Failed to load MSAL4J Java library for performing ActiveDirectoryPassword authentication." 

I believe the first step in fixing this issue is adding MSAL4J as a dependency. I am not a Java developer, so I am out of my comfort zone here if someone more knowledgeable can weigh in.