Performance: Cache instead of recreating Json Validators. (#22060)
Today we recreate JSON validators each time we perform json validation.

Json validation is run twice for each record:

Validation that the record conforms to the general Airbyte Protocol schema.
Validation that the record conforms to that Stream's schema.
Looking at the code, we create at least 3 objects that are discarded for every single record that passes through the platform. This is both CPU and Garbage Collection inefficient. In particular, creating the validator object is expensive since it parses the entire json schema each time.

This CPU/GC inefficiency is true for all code that uses the current Json validator class, which can include Sources and Destinations.

- Instead of recreating the schema validators each time, initialise the validators and reuse them.
- The JsonSchemaValidator class should be rewritten to clean up it's methods and to cache validators by default. I'm skipping this for now to keep the PR small. I'll revisit in a follow up PR.
- Improved the ReplicationWorkerPerformanceTest so the source messages are emitted through the various stream factories for a higher fidelity test.