destination-s3-glue: Add TableType and fix race condition
- The TableType attribute is not getting populated in the Glue catalog which has started
to lead to errors when trying to query the table from e.g. Trino. The error message
observed is `Cannot invoke "String.equals(Object)" because "tableType" is null`
- There is a race condition in initializing the destination connector where a failure
occurs if multiple connections attempt to initialize at the same time, because the test
table that gets created is statically named. This adds a random suffix to the table to
avoid that race condition.
- The Airbyte sync ID and emitted_at fields are useful for deduplicating data. This adds
those fields to the table definition since they are already included in the records as written.