:sparkles: update release pipeline to build platform from airbyte-platform
dev runs of this pr can be found here: https://github.com/airbytehq/airbyte/actions/workflows/release-airbyte-os.yml

I've tested the `create_release` step for both `airbyte` and `airbyte-platform`.
What I haven't tested are the auto commit steps for both.

What also should be noted is the artifacts from the `run-ab-platform.sh` script need to be commited into the repo for them to be included in the release.