Source Zoho CRM: Set airbyte type to string for zoho autonumbers when they include prefix or suffix #23818 (#24034)
* feature: correctly set autonumber type if it contains prefix or suffix

Zoho connector assumes any autonumber data type maps to airbyte's
big_integer. But that is not true as in Zoho you can create incremental
fields with strings in them via the prefix/suffix config params. Taking
this into account we choose string format in case prefix and suffix are
not empty.

* fix unit test and bump connector version

* correct auto_numbe ref

* update docker python base and fix auth

* fix python version file

* fix tests and validate with integration test

* fix  flake

* auto-bump connector version

---------

Co-authored-by: Gerard Clos <clos.gerard@gmail.com>
Co-authored-by: Octavia Squidington III <octavia-squidington-iii@users.noreply.github.com>