Destination S3 Glue: struct fix (#24675)
* destination-s3-glue: Fix struct and array schema generation

The `struct` schema type was being generated with invalid syntax due to the inclusion of
space characters around the `:` character in the type mapping of the struct field.

The `array` data type was not properly handling the optionality of the `items`
property. This lead to hard failures whenever dealing with array types that didn't set
the `items` property. This adds conditional handling of when the `items` type is present
and otherwise assumes that the type is a string.

* rollback change

* bump connector version

* return version to 0.1.3

* auto-bump connector version

---------

Co-authored-by: Tobias Macey <tmacey@mit.edu>
Co-authored-by: Octavia Squidington III <octavia-squidington-iii@users.noreply.github.com>