Update README.md
I noticed several punctuation marks were missing, so I fixed them. I also want to use this opportunity to draw get the attention of the docs team. 
I would love to join Airbyte as a technical writer and be a part of the team. I am more than willing to handle these tasks, even if the organization does not have the current budget to onboard a new technical writer. 


Also, I filled the form to participate in the Airbyte Writer Program. I hope the team reviews my topic and allow me contribute to the data insight and tutorial articles. Cheers!