destination-s3-glue: Fix precision and scale of decimal type
The decimal type in Hive will default to a precision of 10 and a scale of 0, meaning
that it cannot handle numbers larger than 10 digits. The maximum values for precision
and scale are 38 and 38 respectively. This updates the decimal type specified in
destination schemas to use the maximum values for the broadest compatibility with
numeric values.

If a table is using the default precision of 10 and a number is present which exceeds
that many digits then the query engine will throw a `fieldValue cannot be null` error
message which is very cryptic.