Implementation of an ExecutorFactory for async work
The ExecutorFactory just gives us a convenient way to create a ThreadPoolExecutor with the properties we want:
* Configurable maximum task pool size.
* Single core and max thread for serial execution of tasks.
* Default Blocking RejectedExecutionHandler to halt processing until progress can be made.

Note that in this commit, the Executor is not wired in yet. This is just the factory and some tests.

This commit also includes the CloseableResourceManager, a singleton to track long running resources that require closing. Think Executors or database connection pools. A call can be made at process close to close these objects.